delete alert unit pln ok

+++ fe/src/views/UnitPln/UnitPln.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import axios from 'axios';
 import { Card, CardHeader, CardContent, Divider, Grid, Typography } from '@mui/material';
+import Swal from 'sweetalert2';
 
 // project import
 import Breadcrumb from 'component/Breadcrumb';
@@ -21,12 +22,30 @@ export default function UnitPln() {
   };
 
   const deleteUnit = async (id) => {
-    try {
-      await axios.delete(`http://localhost:3000/api/unitpln/${id}`);
-      getUnits();
-    } catch (error) {
-      console.log(error);
-    }
+    // try {
+    //   await axios.delete(`http://localhost:3000/api/unitpln/${id}`);
+    //   getUnits();
+    // } catch (error) {
+    //   console.log(error);
+    // }
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`http://localhost:3000/api/unitpln/${id}`);
+        Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+        setTimeout(() => {
+          getUnits();
+        }, 2000);
+      }
+    });
   };
 
   return (
